Added preliminary recipe input form
The form appears when you select a day/multiple days using the select
event. A drop down box and a button appears. Select something from the
dropdown and press the button. This will close the form and place the
selected data from the dropdown in the calendar.

We need to load the recipes into the dropdown from localstorage. It
should parse them by name. Right now I just have placeholder data for
testing purposes.